Update IDynamicEnumValuesProvider.md
`IDynamicEnumValuesGenerator` has been connected for Update 1. However, we have decided to change the behavior and will be reflecting this in the docs. The new behavior allows for `IDynamicEnumValuesGenerator` the first shot at providing an `IEnumValue` based on the user supplied value. However, if this returns a task of null, CPS will then try to map it into an `IEnumValue` from `GetListedValyesAsync()` on its own. If that does not succeed then the user supplied value will be placed directly into the property field. As such the `AllowCustomValues` field is now obsolete and will be removed eventually.

closes #53