Add link to SDK download
It was not clear to me in reading these docs that this SDK was a separate download. I had installed the "Visual Studio extension development" toolset in the VS installer, but none of the items in these docs were showing up. It was very confusing. Hoping this link will help clarify.