Fix links to blog posts
A previous commit attempted to fix up these links to point to updated
locations on docs.microsoft.com instead of blogs.msdn.com. However, the
updated links actually pointed to unrelated documentation--with a
similar name--on creating a new project system based on MPF (the Managed
Project Framework), rather than the blog posts on creating property
pages in a CPS-based project system. This commit corrects the links.