Document obtaining MEF components from projects
In a chat with a partner team today I was looking for a doc to explain
how to pull a MEF component from a project, when not a suitably-scoped
MEF component yourself.

This information is touched on in `finding_CPS_in_a_VS_project.md`, but
feels more discoverable in `mef.md`.